rag_assignment/requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-community
langchain-openai
sentence-transformers
python-multipart
pypdf
python-dotenv
endee
requests
